refactor cancellation scheduler
