+++ core/src/main/java/com/linecorp/armeria/server/ServiceRequestContextBuilder.java
@@ -15,6 +15,7 @@
  */
 package com.linecorp.armeria.server;
 
+import static com.linecorp.armeria.internal.common.CancellationScheduler.noopCancellationTask;
 import static java.util.Objects.requireNonNull;
 
 import java.net.InetAddress;
@@ -40,14 +41,12 @@ import com.linecorp.armeria.common.logging.RequestLog;
 import com.linecorp.armeria.common.logging.RequestLogBuilder;
 import com.linecorp.armeria.common.util.SystemInfo;
 import com.linecorp.armeria.internal.common.CancellationScheduler;
-import com.linecorp.armeria.internal.common.CancellationScheduler.CancellationTask;
 import com.linecorp.armeria.internal.server.DefaultServiceRequestContext;
 
 import io.micrometer.core.instrument.MeterRegistry;
 import io.netty.buffer.ByteBufAllocator;
 import io.netty.channel.Channel;
 import io.netty.channel.EventLoop;
-import io.netty.util.concurrent.ImmediateEventExecutor;
 
 /**
  * Builds a new {@link ServiceRequestContext}. Note that it is not usually required to create a new context by
@@ -71,26 +70,11 @@ public final class ServiceRequestContextBuilder extends AbstractRequestContextBu
         }
     };
 
-    private static final CancellationTask noopCancellationTask = new CancellationTask() {
-        @Override
-        public boolean canSchedule() {
-            return true;
-        }
-
-        @Override
-        public void run(Throwable cause) { /* no-op */ }
-    };
-
     /**
      * A cancellation scheduler that has been finished.
      */
-    private static final CancellationScheduler noopRequestCancellationScheduler = new CancellationScheduler(0);
-
-    static {
-        noopRequestCancellationScheduler.init(ImmediateEventExecutor.INSTANCE, noopCancellationTask,
-                                              0, /* server */ true);
-        noopRequestCancellationScheduler.finishNow();
-    }
+    private static final CancellationScheduler
+            finishedRequestCancellationScheduler = CancellationScheduler.finished(true);
 
     private final List<Consumer<? super ServerBuilder>> serverConfigurators = new ArrayList<>(4);
 
@@ -254,9 +238,9 @@ public final class ServiceRequestContextBuilder extends AbstractRequestContextBu
 
         final CancellationScheduler requestCancellationScheduler;
         if (timedOut()) {
-            requestCancellationScheduler = noopRequestCancellationScheduler;
+            requestCancellationScheduler = finishedRequestCancellationScheduler;
         } else {
-            requestCancellationScheduler = new CancellationScheduler(0);
+            requestCancellationScheduler = CancellationScheduler.of(0);
             final CountDownLatch latch = new CountDownLatch(1);
             eventLoop().execute(() -> {
                 requestCancellationScheduler.init(eventLoop(), noopCancellationTask, 0, /* server */ true);
